medication_reminder_backend/requirements.txt
# Django Core
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0

# Database
psycopg2-binary==2.9.7
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==0.56.1
cryptography==41.0.4

# API Documentation
drf-spectacular==0.26.4

# Task Queue & Background Jobs
celery==5.3.1
redis==4.6.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.0.0

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
pytest-django==4.5.2
pytest==7.4.0

# Production
gunicorn==21.2.0
whitenoise==6.5.0
